+++ deepx/node/__init__.py
@@ -1,4 +1,4 @@
 from .node import Node
 from .data import Data
-from .sequence import RecurrentNode, Sequence, Generate
 from .exceptions import *
+from .rnn import *
