+++ deepx/rnn/__init__.py
@@ -1,3 +1,3 @@
 from .data import *
 from .lstm import *
-from ..node import Sequence, Generate
+from .generate import *
